+++ test/flycheck-test.el
@@ -3306,46 +3306,39 @@ of the file will be interrupted because there are too many #ifdef configurations
 
 (ert-deftest flycheck-d-base-directory/no-module-declaration ()
   :tags '(builtin-checker language-d)
-  (flycheck-test-with-resource-buffer "checkers/d_dmd_syntax_error_without_module.d"
+  (flycheck-test-with-resource-buffer "checkers/d/src/dmd/no_module.d"
     (should (f-same? (flycheck-d-base-directory)
-                     (f-join flycheck-test-resources-directory "checkers")))))
+                     (flycheck-test-resource-filename "checkers/d/src/dmd")))))
 
 (ert-deftest flycheck-d-base-directory/with-module-declaration ()
   :tags '(builtin-checker language-d)
-  (flycheck-test-with-resource-buffer "checkers/d_dmd_warning.d"
+  (flycheck-test-with-resource-buffer "checkers/d/src/dmd/warning.d"
     (should (f-same? (flycheck-d-base-directory)
-                     flycheck-test-resources-directory))))
+                     (flycheck-test-resource-filename "checkers/d/src")))))
 
 (ert-deftest flycheck-d-base-directory/package-file ()
   :tags '(builtin-checker language-d)
-  (flycheck-test-with-resource-buffer "checkers/package.d"
+  (flycheck-test-with-resource-buffer "checkers/d/src/dmd/package.d"
     (should (f-same? (flycheck-d-base-directory)
-                     flycheck-test-resources-directory))))
+                     (flycheck-test-resource-filename "checkers/d/src")))))
 
-(ert-deftest flycheck-define-checker/d-dmd-syntax-error ()
+(ert-deftest flycheck-define-checker/d-dmd-warning-include-path ()
+  :tags '(builtin-checker external-tool language-d)
+  (skip-unless (flycheck-check-executable 'd-dmd))
+  (let ((flycheck-dmd-include-path '("../../lib")))
+    (flycheck-test-should-syntax-check
+     "checkers/d/src/dmd/warning.d" 'd-mode
+     '(9 nil warning "statement is not reachable" :checker d-dmd)
+     '(20 nil warning "function dmd.warning.bar is deprecated"
+          :checker d-dmd))))
+
+(ert-deftest flycheck-define-checker/d-dmd-missing-import ()
   :tags '(builtin-checker external-tool language-d)
   (skip-unless (flycheck-check-executable 'd-dmd))
   (flycheck-test-should-syntax-check
-   "checkers/d-dmd-syntax-error.d" 'd-mode
-   '(2 nil error "module studio is in file 'std/studio.d' which cannot be read"
+   "checkers/d/src/dmd/warning.d" 'd-mode
+   '(4 nil error "module external_library is in file 'external_library.d' which cannot be read"
        :checker d-dmd)))
-
-(ert-deftest flycheck-define-checker/d-dmd-syntax-error-without-module ()
-  :tags '(builtin-checker external-tool language-d)
-  (skip-unless (flycheck-check-executable 'd-dmd))
-  (flycheck-test-should-syntax-check
-   "checkers/d_dmd_syntax_error_without_module.d" 'd-mode
-   '(5 nil error "undefined identifier writel, did you mean template write(T...)(T args) if (!is(T[0] : File))?"
-       :checker d-dmd)))
-
-(ert-deftest flycheck-define-checker/d-dmd-warning ()
-  :tags '(builtin-checker external-tool language-d)
-  (skip-unless (flycheck-check-executable 'd-dmd))
-  (flycheck-test-should-syntax-check
-   "checkers/d_dmd_warning.d" 'd-mode
-   '(6 nil warning "statement is not reachable" :checker d-dmd)
-   '(17 nil warning "function checkers.d_dmd_warning.bar is deprecated"
-        :checker d-dmd)))
 
 (ert-deftest flycheck-define-checker/elixir-error ()
   :tags '(builtin-checker external-tool language-elixir)