+++ test/resources/checkers/d-dmd-syntax-error.d
@@ -1,7 +1,0 @@
-module d_dmd_syntax_error;
-import std.studio;
-
-void main()
-{
-    writeln("Hello, world!");
-}