+++ test/resources/checkers/d/src/dmd/library.d
@@ -1,0 +1,3 @@
+module dmd.library;
+
+// A in-source library, to test the implicit inclusion of the base directory.