+++ test/resources/checkers/d/src/dmd/package.d
@@ -1,3 +1,3 @@
 // A dummy D package file to test base calculation
 
-module checkers;
+module dmd;