+++ test/resources/checkers/d/src/dmd/warning.d
@@ -1,4 +1,7 @@
-module checkers.d_dmd_warning;
+module dmd.warning;
+
+import dmd.library;
+import external_library;
 
 auto foo(int a)
 {