+++ test/resources/checkers/d_dmd_syntax_error_without_module.d
@@ -1,6 +1,0 @@
-import std.stdio;
-
-void main()
-{
-    writel("Hello");
-}